criado os dtos de listar e cadastrar

+++ src/main/java/com/api/sgpcbackend/domain/dto/patrimonio/ComputadorCadastroDTO.java
@@ -4,29 +4,15 @@ import com.api.sgpcbackend.domain.model.patrimonio.computador.espec.*;
 
 public record ComputadorCadastroDTO(
         String tombamento,
-
         String descricao,
-
-        Integer estado,
-
-        //Integer tipo,
-
-        Integer localidade,
-
+        Short estado,
+        Short localidade,
         Boolean alienado,
-
         String serial,
-
-        Integer modelo,
-
-        Integer sistema_operacional,
-
-        Integer ram,
-
-        Integer ram_ddr,
-
-        Integer hd
+        Short modelo,
+        Short sistema_operacional,
+        Short ram,
+        Short ram_ddr,
+        Short hd
 )
-{
-
-}
+{}
