+++ src/main/java/com/api/sgpcbackend/domain/dto/patrimonio/PatrimonioCadastroDTO.java
@@ -2,19 +2,11 @@ package com.api.sgpcbackend.domain.dto.patrimonio;
 
 
 public record PatrimonioCadastroDTO(
-
      String tombamento,
-
      String descricao,
-
-     Integer estado,
-
-     Integer tipo,
-
-     Integer localidade,
-
+     Short estado,
+     Short tipo,
+     Short localidade,
      Boolean alienado
-
     )
-{
-}
+{}
