+++ src/main/java/com/api/sgpcbackend/domain/model/patrimonio/computador/espec/HD.java
@@ -15,12 +15,12 @@ public class HD
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Short id;
 
     @Column(name = "descricao")
     private String descricao;
 
-    public HD(Integer id)
+    public HD(Short id)
     {
         this.id = id;
     }
